# Lab2/pzpi-22-3-shpak-denys-lab2/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateSensorDataRequest:
    properties:
      date:
        description: |-
          Date is the date and time when the sensor data was recorded
          example: 2021-09-01T12:00:00Z
        example: "2021-09-01T12:00:00Z"
        type: string
      humidity:
        description: |-
          Humidity is the humidity recorded by the sensor
          example: 50.0
        example: 50
        type: number
      pressure:
        description: |-
          MeanPressure is the mean pressure recorded by the sensor
          example: 1013.25
        example: 1013.25
        type: number
      temperature:
        description: |-
          Temperature is the temperature recorded by the sensor
          example: 25.5
        example: 25.5
        type: number
      waypoint_id:
        description: |-
          WaypointID is the ID of the waypoint where the sensor data was recorded
          example: 1
        type: integer
      wind_speed:
        description: |-
          WindSpeed is the wind speed recorded by the sensor
          example: 10.5
        example: 10.5
        type: number
    type: object
...
  handlers.UpdateSensorDataRequest:
    properties:
      date:
        description: |-
          Date is the date and time when the sensor data was recorded
          example: 2021-09-01T12:00:00Z
        example: "2021-09-01T12:00:00Z"
        type: string
      humidity:
        description: |-
          Humidity is the humidity recorded by the sensor
          example: 50.0
        example: 50
        type: number
      temperature:
        description: |-
          Temperature is the temperature recorded by the sensor
          example: 25.5
        example: 25.5
        type: number
    type: object
...
  handlers.UpdateUserRequest:
    properties:
      name:
        description: |-
          Name is the name of the user
          Example: John Doe
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
...
  /sensor-data/:
    post:
      consumes:
      - application/json
      description: Adds new sensor data to the specified SensorData
      parameters:
      - description: Sensor data details
        in: body
        name: sensor_data
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateSensorDataRequest'
      produces:
      - application/json
      responses: {}
      summary: Add sensor data to a SensorData
      tags:
      - sensor
  /sensor-data/{sensor_data_id}:
    delete:
      description: Deletes sensor data with the given ID
      parameters:
      - description: Sensor Data ID
        in: path
        name: sensor_data_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete sensor data by ID
      tags:
      - sensor
    get:
      description: Retrieves sensor data with the given ID
      parameters:
      - description: Sensor Data ID
        in: path
        name: sensor_data_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: Get sensor data by ID
      tags:
      - sensor
    put:
      consumes:
      - application/json
      description: Updates sensor data with the given ID
      parameters:
      - description: Sensor Data ID
        in: path
        name: sensor_data_id
        required: true
        type: integer
      - description: Sensor data details
        in: body
        name: sensor_data
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSensorDataRequest'
      produces:
      - application/json
      responses: {}
      summary: Update sensor data by ID
      tags:
      - sensor
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user by ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a user by ID
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      operationId: get-users
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get users
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the Bearer token in the format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"